Extract usage of `native-keymap` into IKeyboardLayoutService
