+++ src/vs/workbench/services/keybinding/common/keymapInfo.ts
@@ -9,87 +9,9 @@ import { createDecorator } from 'vs/platform/instantiation/common/instantiation'
 import { DispatchConfig } from 'vs/workbench/services/keybinding/common/dispatchConfig';
 import { IKeyboardMapper } from 'vs/workbench/services/keybinding/common/keyboardMapper';
 import { IKeyboardEvent } from 'vs/platform/keybinding/common/keybinding';
+import * as keyboardLayout from 'vs/workbench/services/keyboardLayout/common/keyboardLayout';
 
-
-export interface IWindowsKeyMapping {
-	vkey: string;
-	value: string;
-	withShift: string;
-	withAltGr: string;
-	withShiftAltGr: string;
-}
-export interface IWindowsKeyboardMapping {
-	[code: string]: IWindowsKeyMapping;
-}
-export interface ILinuxKeyMapping {
-	value: string;
-	withShift: string;
-	withAltGr: string;
-	withShiftAltGr: string;
-}
-export interface ILinuxKeyboardMapping {
-	[code: string]: ILinuxKeyMapping;
-}
-export interface IMacKeyMapping {
-	value: string;
-	withShift: string;
-	withAltGr: string;
-	withShiftAltGr: string;
-	valueIsDeadKey: boolean;
-	withShiftIsDeadKey: boolean;
-	withAltGrIsDeadKey: boolean;
-	withShiftAltGrIsDeadKey: boolean;
-}
-export interface IMacKeyboardMapping {
-	[code: string]: IMacKeyMapping;
-}
-
-export type IKeyboardMapping = IWindowsKeyboardMapping | ILinuxKeyboardMapping | IMacKeyboardMapping;
-
-/* __GDPR__FRAGMENT__
-	"IKeyboardLayoutInfo" : {
-		"name" : { "classification": "SystemMetaData", "purpose": "FeatureInsight" },
-		"id": { "classification": "SystemMetaData", "purpose": "FeatureInsight" },
-		"text": { "classification": "SystemMetaData", "purpose": "FeatureInsight" }
-	}
-*/
-export interface IWindowsKeyboardLayoutInfo {
-	name: string;
-	id: string;
-	text: string;
-}
-
-/* __GDPR__FRAGMENT__
-	"IKeyboardLayoutInfo" : {
-		"model" : { "classification": "SystemMetaData", "purpose": "FeatureInsight" },
-		"layout": { "classification": "SystemMetaData", "purpose": "FeatureInsight" },
-		"variant": { "classification": "SystemMetaData", "purpose": "FeatureInsight" },
-		"options": { "classification": "SystemMetaData", "purpose": "FeatureInsight" },
-		"rules": { "classification": "SystemMetaData", "purpose": "FeatureInsight" }
-	}
-*/
-export interface ILinuxKeyboardLayoutInfo {
-	model: string;
-	layout: string;
-	variant: string;
-	options: string;
-	rules: string;
-}
-
-/* __GDPR__FRAGMENT__
-	"IKeyboardLayoutInfo" : {
-		"id" : { "classification": "SystemMetaData", "purpose": "FeatureInsight" },
-		"lang": { "classification": "SystemMetaData", "purpose": "FeatureInsight" },
-		"localizedName": { "classification": "SystemMetaData", "purpose": "FeatureInsight" }
-	}
-*/
-export interface IMacKeyboardLayoutInfo {
-	id: string;
-	lang: string;
-	localizedName?: string;
-}
-
-export type IKeyboardLayoutInfo = (IWindowsKeyboardLayoutInfo | ILinuxKeyboardLayoutInfo | IMacKeyboardLayoutInfo) & { isUserKeyboardLayout?: boolean; isUSStandard?: true };
+export type IKeyboardLayoutInfo = keyboardLayout.IKeyboardLayoutInfo & { isUserKeyboardLayout?: boolean; isUSStandard?: true };
 
 export const IKeymapService = createDecorator<IKeymapService>('keymapService');
 
@@ -99,94 +21,8 @@ export interface IKeymapService {
 	getKeyboardMapper(dispatchConfig: DispatchConfig): IKeyboardMapper;
 	getCurrentKeyboardLayout(): IKeyboardLayoutInfo | null;
 	getAllKeyboardLayouts(): IKeyboardLayoutInfo[];
-	getRawKeyboardMapping(): IKeyboardMapping | null;
+	getRawKeyboardMapping(): keyboardLayout.IKeyboardMapping | null;
 	validateCurrentKeyboardMapping(keyboardEvent: IKeyboardEvent): void;
-}
-
-export function areKeyboardLayoutsEqual(a: IKeyboardLayoutInfo | null, b: IKeyboardLayoutInfo | null): boolean {
-	if (!a || !b) {
-		return false;
-	}
-
-	if ((<IWindowsKeyboardLayoutInfo>a).name && (<IWindowsKeyboardLayoutInfo>b).name && (<IWindowsKeyboardLayoutInfo>a).name === (<IWindowsKeyboardLayoutInfo>b).name) {
-		return true;
-	}
-
-	if ((<IMacKeyboardLayoutInfo>a).id && (<IMacKeyboardLayoutInfo>b).id && (<IMacKeyboardLayoutInfo>a).id === (<IMacKeyboardLayoutInfo>b).id) {
-		return true;
-	}
-
-	if ((<ILinuxKeyboardLayoutInfo>a).model &&
-		(<ILinuxKeyboardLayoutInfo>b).model &&
-		(<ILinuxKeyboardLayoutInfo>a).model === (<ILinuxKeyboardLayoutInfo>b).model &&
-		(<ILinuxKeyboardLayoutInfo>a).layout === (<ILinuxKeyboardLayoutInfo>b).layout
-	) {
-		return true;
-	}
-
-	return false;
-}
-
-export function parseKeyboardLayoutDescription(layout: IKeyboardLayoutInfo | null): { label: string, description: string } {
-	if (!layout) {
-		return { label: '', description: '' };
-	}
-
-	if ((<IWindowsKeyboardLayoutInfo>layout).name) {
-		// windows
-		let windowsLayout = <IWindowsKeyboardLayoutInfo>layout;
-		return {
-			label: windowsLayout.text,
-			description: ''
-		};
-	}
-
-	if ((<IMacKeyboardLayoutInfo>layout).id) {
-		let macLayout = <IMacKeyboardLayoutInfo>layout;
-		if (macLayout.localizedName) {
-			return {
-				label: macLayout.localizedName,
-				description: ''
-			};
-		}
-
-		if (/^com\.apple\.keylayout\./.test(macLayout.id)) {
-			return {
-				label: macLayout.id.replace(/^com\.apple\.keylayout\./, '').replace(/-/, ' '),
-				description: ''
-			};
-		}
-		if (/^.*inputmethod\./.test(macLayout.id)) {
-			return {
-				label: macLayout.id.replace(/^.*inputmethod\./, '').replace(/[-\.]/, ' '),
-				description: `Input Method (${macLayout.lang})`
-			};
-		}
-
-		return {
-			label: macLayout.lang,
-			description: ''
-		};
-	}
-
-	let linuxLayout = <ILinuxKeyboardLayoutInfo>layout;
-
-	return {
-		label: linuxLayout.layout,
-		description: ''
-	};
-}
-
-export function getKeyboardLayoutId(layout: IKeyboardLayoutInfo): string {
-	if ((<IWindowsKeyboardLayoutInfo>layout).name) {
-		return (<IWindowsKeyboardLayoutInfo>layout).name;
-	}
-
-	if ((<IMacKeyboardLayoutInfo>layout).id) {
-		return (<IMacKeyboardLayoutInfo>layout).id;
-	}
-
-	return (<ILinuxKeyboardLayoutInfo>layout).layout;
 }
 
 function deserializeMapping(serializedMapping: ISerializedMapping) {
@@ -313,7 +149,7 @@ export class KeymapInfo {
 			return false;
 		}
 
-		if (getKeyboardLayoutId(this.layout) !== getKeyboardLayoutId(other.layout)) {
+		if (keyboardLayout.getKeyboardLayoutId(this.layout) !== keyboardLayout.getKeyboardLayoutId(other.layout)) {
 			return false;
 		}
