+++ src/vs/workbench/workbench.desktop.main.ts
@@ -67,6 +67,7 @@ import 'vs/workbench/electron-browser/desktop.main';
 import 'vs/workbench/services/integrity/node/integrityService';
 import 'vs/workbench/services/search/electron-browser/searchService';
 import 'vs/workbench/services/textfile/electron-browser/nativeTextFileService';
+import 'vs/workbench/services/keyboardLayout/electron-browser/keyboardLayout';
 import 'vs/workbench/services/keybinding/electron-browser/nativeKeymapService';
 import 'vs/workbench/services/extensions/electron-browser/extensionService';
 import 'vs/workbench/services/extensionManagement/electron-browser/extensionManagementServerService';
